+++ NEWS.md
@@ -5,6 +5,7 @@
 - Added function to mock-up package creation so examples can run for CRAN submission.
 - Added `npm_uninstall` to uninstall NPM packages.
 - Deprecate `bundle` in favour of `bund_dev` and `bundle_prod` as `bundle` is confusing and error prone.
+- Added `npm_update` and `npm_outdated`.
 
 ## packer 0.0.5